+++ src/main/java/com/example/wineapi/data/dto/question/MultipleChoiceQuestionDto.java
@@ -1,0 +1,22 @@
+package com.example.wineapi.data.dto.question;
+
+import com.example.wineapi.data.entity.question.Question;
+
+import java.util.ArrayList;
+
+public class MultipleChoiceQuestionDto extends QuestionDto{
+    private ArrayList<String> question_option;
+
+    public MultipleChoiceQuestionDto(Question question) {
+        super(question);
+        this.question_option = new ArrayList<>();
+    }
+
+    public ArrayList<String> getQuestion_option() {
+        return question_option;
+    }
+
+    public void setQuestion_option(ArrayList<String> question_option) {
+        this.question_option = question_option;
+    }
+}
